feat: extract moderator escalation intervention into separate moderatorNotifier agent
allow logic to escalate to moderator based on shared and private chat to run independently/more
often

BREAKING CHANGE: removes eventMediatorPlus agent, impacts only active conversations
