+++ tests/agents/helpers/interventionHandler.test.ts
@@ -16,14 +16,13 @@ describe('interventionHandler', () => {
       expect(InterventionType.BRIDGE).toBe('BRIDGE')
       expect(InterventionType.STRUCTURE).toBe('STRUCTURE')
       expect(InterventionType.PLAY).toBe('PLAY')
-      expect(InterventionType.MODERATOR_ESCALATION).toBe('MODERATOR_ESCALATION')
       expect(InterventionType.NONE).toBe('NONE')
     })
   })
 
   describe('getInterventionAnalysisSchema', () => {
     const allInterventions = Object.values(InterventionType)
-    const schema = getInterventionAnalysisSchema(allInterventions, true)
+    const schema = getInterventionAnalysisSchema(allInterventions)
 
     it('validates a valid intervention analysis with intervention', () => {
       const validAnalysis = {
@@ -49,23 +48,6 @@ describe('interventionHandler', () => {
         interventionType: 'NONE',
         reasoning: 'No patterns detected, waiting for more signals',
         confidenceScore: 30
-      }
-
-      const result = schema.safeParse(validAnalysis)
-      expect(result.success).toBe(true)
-    })
-
-    it('validates moderator escalation type', () => {
-      const validAnalysis = {
-        shouldIntervene: true,
-        interventionType: 'MODERATOR_ESCALATION',
-        reasoning: 'Strong recurring theme needs moderator attention',
-        sharedChatMessage: "I've flagged a question to the moderator",
-        moderatorMessage:
-          'Multiple participants asking about regulatory changes. Suggested question: How do recent regulatory shifts affect adoption?',
-        confidenceScore: 90,
-        detectedPattern: 'Regulatory concerns',
-        affectedUsers: 6
       }
 
       const result = schema.safeParse(validAnalysis)
@@ -129,7 +111,6 @@ describe('interventionHandler', () => {
         'BRIDGE',
         'STRUCTURE',
         'PLAY',
-        'MODERATOR_ESCALATION',
         'NONE'
       ]
 
@@ -149,7 +130,7 @@ describe('interventionHandler', () => {
     it('only allows enabled intervention types in schema', () => {
       // Create schema with only engagement interventions
       const engagementOnly = [InterventionType.PROVOCATION, InterventionType.PLAY, InterventionType.NONE]
-      const limitedSchema = getInterventionAnalysisSchema(engagementOnly, false)
+      const limitedSchema = getInterventionAnalysisSchema(engagementOnly)
 
       // Should accept enabled types
       const validAnalysis = {
@@ -168,38 +149,6 @@ describe('interventionHandler', () => {
         confidenceScore: 75
       }
       expect(limitedSchema.safeParse(invalidAnalysis).success).toBe(false)
-    })
-
-    it('includes moderatorMessage field when supportsModerator is true', () => {
-      const schemaWithModerator = getInterventionAnalysisSchema(allInterventions, true)
-
-      const analysisWithModeratorMessage = {
-        shouldIntervene: true,
-        interventionType: 'MODERATOR_ESCALATION',
-        reasoning: 'Test',
-        moderatorMessage: 'Test moderator message',
-        confidenceScore: 75
-      }
-
-      const result = schemaWithModerator.safeParse(analysisWithModeratorMessage)
-      expect(result.success).toBe(true)
-    })
-
-    it('schema works without moderatorMessage field when supportsModerator is false', () => {
-      const schemaWithoutModerator = getInterventionAnalysisSchema(
-        [InterventionType.PROVOCATION, InterventionType.NONE],
-        false
-      )
-
-      const analysisWithoutModeratorMessage = {
-        shouldIntervene: true,
-        interventionType: 'PROVOCATION',
-        reasoning: 'Test',
-        confidenceScore: 75
-      }
-
-      const result = schemaWithoutModerator.safeParse(analysisWithoutModeratorMessage)
-      expect(result.success).toBe(true)
     })
   })
 
@@ -221,7 +170,6 @@ describe('interventionHandler', () => {
       expect(varNames).toContain('retrievedChunks')
       expect(varNames).toContain('privateMessages')
       expect(varNames).toContain('sharedChatHistory')
-      expect(varNames).toContain('moderatorContext')
       expect(varNames).toContain('agentRecentPosts')
     })
   })
@@ -233,7 +181,6 @@ describe('interventionHandler', () => {
       expect(USER_TEMPLATE).toContain('## Retrieved Relevant Context')
       expect(USER_TEMPLATE).toContain('## Private Messages')
       expect(USER_TEMPLATE).toContain('## Shared Chat History:')
-      expect(USER_TEMPLATE).toContain('## Moderator Context:')
       expect(USER_TEMPLATE).toContain('## Your Recent Posts:')
     })
 
@@ -243,7 +190,6 @@ describe('interventionHandler', () => {
       expect(USER_TEMPLATE).toContain('{retrievedChunks}')
       expect(USER_TEMPLATE).toContain('{privateMessages}')
       expect(USER_TEMPLATE).toContain('{sharedChatHistory}')
-      expect(USER_TEMPLATE).toContain('{moderatorContext}')
       expect(USER_TEMPLATE).toContain('{agentRecentPosts}')
     })
   })
